:rocket: Create Subpage Arch. for live/:id page
