+++ web/config/i18n/messages/ja.json
@@ -193,9 +193,29 @@
     "youtube": {
       "live": {
         "id": {
-          "metadata": {
-            "title": "{channel} {title}のスパチャ金額、同接数",
-            "description": "YouTubeライブを視聴！{channel}のスパチャ金額や同接数をチェックし人気の瞬間を見逃さず確認しよう。PeakXではなんとスパチャコメントもすべて掲載！VTuberファン必見の情報をお届けします。"
+          "index": {
+            "metadata": {
+              "title": "{channel} {title}のスパチャ金額、同接数",
+              "description": "YouTubeライブを視聴！{channel}のスパチャ金額や同接数をチェックし人気の瞬間を見逃さず確認しよう。PeakXではなんとスパチャコメントもすべて掲載！VTuberファン必見の情報をお届けします。"
+            }
+          },
+          "superChat": {
+            "metadata": {
+              "title": "{channel} {title}のスパチャ金額、同接数",
+              "description": "YouTubeライブを視聴！{channel}のスパチャ金額や同接数をチェックし人気の瞬間を見逃さず確認しよう。PeakXではなんとスパチャコメントもすべて掲載！VTuberファン必見の情報をお届けします。"
+            },
+            "comments": {
+              "metadata": {
+                "title": "{channel} {title}のスパチャ金額、同接数",
+                "description": "YouTubeライブを視聴！{channel}のスパチャ金額や同接数をチェックし人気の瞬間を見逃さず確認しよう。PeakXではなんとスパチャコメントもすべて掲載！VTuberファン必見の情報をお届けします。"
+              }
+            }
+          },
+          "comments": {
+            "metadata": {
+              "title": "{channel} {title}のスパチャ金額、同接数",
+              "description": "YouTubeライブを視聴！{channel}のスパチャ金額や同接数をチェックし人気の瞬間を見逃さず確認しよう。PeakXではなんとスパチャコメントもすべて掲載！VTuberファン必見の情報をお届けします。"
+            }
           },
           "button": {
             "theaterMode": "シアターモードで大きく表示",
